Fixing bugs

+++ hooks/pre_gen.dart
@@ -6,12 +6,19 @@ const desktopPlatforms = "windows,linux,macos";
 const webPlatform = "web";
 
 Future<void> run(HookContext context) async {
+  context.vars["useEmptySample"]=false;
+
+  // Validation app name
+  if(context.vars["projectName"].contains(RegExp(r'[A-Z]'))) {
+    return context.logger.err("App name must consist of lowercase, numbers or underscore characters (e.g. counter_app1).");
+  }
+
   // Add an additional question if counterExample false
-  String sampleId = "custom.emptyApp.1";
+  String? sampleId;
   if (context.vars["counterExample"] == false) {
     sampleId = await context.logger.prompt(
-        "? Want to use a different Flutter Sample instead? (Enter ID from https://samples.flutter.de)",
-        defaultValue: "custom.emptyApp.1");
+        "? Want to use a Flutter Sample instead? (Enter ID from https://samples.flutter.de)",
+        defaultValue: null);
   }
 
   // Define platforms
@@ -28,12 +35,12 @@ Future<void> run(HookContext context) async {
     'create',
     '--suppress-analytics',
     '--org',
-    '{{packageName}}',
+    '{{orgaName}}',
     '{{projectName}}',
     '--platforms',
     platforms
   ];
-  if (sampleId.isNotEmpty && !sampleId.startsWith("custom")) {
+  if (sampleId!=null && sampleId.isNotEmpty && !sampleId.startsWith("custom")) {
     args = [...args, '--sample', sampleId];
     context.logger.info('Creating your flutter app with sample (${sampleId})...');
   } else {
@@ -43,16 +50,11 @@ Future<void> run(HookContext context) async {
   stdout.write(result.stdout);
   stderr.write(result.stderr);
 
-  // Optional: Remove Flutter counter example
-  if (context.vars["counterExample"] == false) {
+  // Optional: Remove Flutter counter example & use our default
+  if (context.vars["counterExample"] == false && sampleId==null) {
     var currentDir = Directory.current;
     final projectPath = "${currentDir.path}/${context.vars["projectName"]}";
     await File('$projectPath/lib/main.dart').delete();
-
-    if(sampleId == "custom.emptyApp.1") {
-      context.vars["useEmptySample"] = true;
-    } else {
-      
-    }
+    context.vars["useEmptySample"]=true;
   }
 }
